// ExistingCode/src/data-access/domain/tables/geologycombinedlog/schema.ts
export const GeologyCombinedLogTableInsertBaseSchema = z.object({
// ...
	comments: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	quickLogInd: z.any().optional(),
	/**
	 * dataType: varbinary
	 */
	migrationRv: z.any().nullable().optional(),
	/**
	 * dataType: int
	 */
	migrationID: z.number().int().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	jsonData: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	reportIncludeInd: z.any().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	validationStatus: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 */
	validationErrors: z.string().max(4000).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	supersededById: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (CONVERT([nvarchar](50),coalesce(session_context(N'UserName'),suser_sname())))
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});